src/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Markdown==3.6
coreapi==2.3.3
psycopg2-binary==2.9.9